Fix for race condition for final file
The final m4b file creation throws a read error on the temp directory.  This appears to be due to a race condition in the create m4b function.  This change fixes the issue though it does keep the temp folders around when the application finishes but the job finishes without erroring.